Include scsi_wait_scan in initramfs (chaos b1294)
Add an 'installkernel' script to the 'xnfsroot' dracut module
whose job is to install scsi_wait_scan in the initramfs.
scsi_wait_scan is modprobed in base/init when udev takes too
long to bring up the network.  It probably should be pulled in by
base; instead it is pulled in by kernel-modules which nfsroot disables.